requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.26.0

# Database connectivity
pyodbc>=5.0.0
sqlalchemy>=2.0.0

# Dashboard
streamlit>=1.32.0
plotly>=5.20.0

# AI Assistant
google-genai>=1.0.0
groq>=1.0.0

# External APIs
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Excel support (optional exports)
openpyxl>=3.1.0

# Testing
pytest>=8.0.0
